+++ CHANGELOG.md
@@ -18,6 +18,7 @@ breaking changes compared to the pre-release version.
   parameter.
 - Support for transfers to S3 to `getEvents()`
 - `PER_FILE` compensation constant
+- Support for `geometricMean` as a statistics value in `getStatistics()`.
 
 ### Changed
 - **Breaking** Failed API requests will now stop instead of warn.